+++ sangmin/java-mid1/src/main/java/section6/Member.java
@@ -27,7 +27,7 @@ public class Member {
         return account.getAccountNumber();
     }
 
-    public int getBalance() {
+    public long getBalance() {
         return account.getBalance();
     }
