+++ sangmin/java-mid1/src/main/java/section6/Transaction.java
@@ -1,5 +1,5 @@
 package section6;
 
 public interface Transaction {
-    int execute(Atm atm, int amount, long targetAccountNumber);
+    long execute(Atm atm, int amount, long targetAccountNumber);
 }
