+++ sangmin/java-mid1/src/main/java/section6/transaction/Withdraw.java
@@ -5,7 +5,7 @@ import section6.Transaction;
 
 public class Withdraw implements Transaction {
     @Override
-    public int execute(Atm atm, int amount, long targetAccountNumber) {
+    public long execute(Atm atm, int amount, long targetAccountNumber) {
         return atm.withdraw(targetAccountNumber, amount);
     }
 }
